chore: write swagger api spec
