+++ internal/catalog/attribute_category_handler.go
@@ -11,6 +11,17 @@ type attributeCategoryHandler struct {
 	service IAttributeService
 }
 
+// categories godoc
+// @Schemes
+// @Summary Categories List
+// @Description Get Categories List.
+// @Tags Attribute Categories
+// @Accept json
+// @Produce json
+// @Success 200 {object} utils.SuccessRespond{data=[]model.Category} "OK RESPOND"
+// @Failure 401 {object} utils.ErrorRespond "UNAUTHORIZED RESPOND"
+// @Failure 500 {object} utils.ErrorRespond "INTERNAL SERVER ERROR RESPOND"
+// @Router /api/v1/categories [GET]
 func (handler attributeCategoryHandler) fetch(ctx *gin.Context) {
 	data, err := handler.service.CategoryList(ctx)
 	if err != nil {
@@ -21,6 +32,19 @@ func (handler attributeCategoryHandler) fetch(ctx *gin.Context) {
 	utils.NewHTTPRespond(ctx, http.StatusOK, data)
 }
 
+// categories godoc
+// @Schemes
+// @Summary Store Category Data
+// @Description Create new Category.
+// @Tags Attribute Categories
+// @Accept json
+// @Produce json
+// @Param body formData NewCategoryForm true "body"
+// @Success 201 {object} utils.SuccessRespond{data=model.Category} "CREATED RESPOND"
+// @Failure 401 {object} utils.ErrorRespond "UNAUTHORIZED RESPOND"
+// @Failure 422 {object} utils.ValidationErrorRespond "UNPROCESSABLE ENTITY RESPOND"
+// @Failure 500 {object} utils.ErrorRespond "INTERNAL SERVER ERROR RESPOND"
+// @Router /api/v1/categories [POST]
 func (handler attributeCategoryHandler) add(ctx *gin.Context) {
 	var form NewCategoryForm
 
@@ -44,6 +68,20 @@ func (handler attributeCategoryHandler) add(ctx *gin.Context) {
 	utils.NewHTTPRespond(ctx, http.StatusCreated, nil)
 }
 
+// categories godoc
+// @Schemes
+// @Summary Update Category Data
+// @Description Update Category Data by ID.
+// @Tags Attribute Categories
+// @Accept json
+// @Produce json
+// @Param id path int true "category id"
+// @Param body formData EditCategoryForm true "body"
+// @Failure 400 {object} utils.ErrorRespond "BAD REQUEST RESPOND"
+// @Failure 401 {object} utils.ErrorRespond "UNAUTHORIZED RESPOND"
+// @Failure 422 {object} utils.ValidationErrorRespond "UNPROCESSABLE ENTITY RESPOND"
+// @Failure 500 {object} utils.ErrorRespond "INTERNAL SERVER ERROR RESPOND"
+// @Router /api/v1/categories/{id} [PATCH]
 func (handler attributeCategoryHandler) edit(ctx *gin.Context) {
 	id, ok := utils.GetIDParam(ctx, "id")
 	if !ok {
@@ -73,6 +111,19 @@ func (handler attributeCategoryHandler) edit(ctx *gin.Context) {
 	utils.NewHTTPRespond(ctx, http.StatusOK, nil)
 }
 
+// categories godoc
+// @Schemes
+// @Summary Delete Category Data
+// @Description Delete Category Data by ID.
+// @Tags Attribute Categories
+// @Accept json
+// @Produce json
+// @Param id path int true "category id"
+// @Success 204 "NO CONTENT RESPOND"
+// @Failure 400 {object} utils.ErrorRespond "BAD REQUEST RESPOND"
+// @Failure 401 {object} utils.ErrorRespond "UNAUTHORIZED RESPOND"
+// @Failure 500 {object} utils.ErrorRespond "INTERNAL SERVER ERROR RESPOND"
+// @Router /api/v1/categories/{id} [DELETE]
 func (handler attributeCategoryHandler) destroy(ctx *gin.Context) {
 	id, ok := utils.GetIDParam(ctx, "id")
 	if !ok {
@@ -87,6 +138,20 @@ func (handler attributeCategoryHandler) destroy(ctx *gin.Context) {
 	utils.NewHTTPRespond(ctx, http.StatusNoContent, nil)
 }
 
+// categories godoc
+// @Schemes
+// @Summary Delete Subcategory in Category Data
+// @Description Delete Subcategory in Category Data by ID.
+// @Tags Attribute Categories
+// @Accept json
+// @Produce json
+// @Param id path int true "category id"
+// @Param sid path int true "subcategory id"
+// @Success 204 "NO CONTENT RESPOND"
+// @Failure 400 {object} utils.ErrorRespond "BAD REQUEST RESPOND"
+// @Failure 401 {object} utils.ErrorRespond "UNAUTHORIZED RESPOND"
+// @Failure 500 {object} utils.ErrorRespond "INTERNAL SERVER ERROR RESPOND"
+// @Router /api/v1/categories/{id}/subcategories/{sid} [DELETE]
 func (handler attributeCategoryHandler) destroySub(ctx *gin.Context) {
 	cid, ok := utils.GetIDParam(ctx, "id")
 	if !ok {
