+++ internal/catalog/attribute_unit_handler.go
@@ -11,6 +11,17 @@ type attributeUnitHandler struct {
 	service IAttributeService
 }
 
+// units godoc
+// @Schemes
+// @Summary Units List
+// @Description Get Units List.
+// @Tags Attribute Units
+// @Accept json
+// @Produce json
+// @Success 200 {object} utils.SuccessRespond{data=[]model.Unit} "OK RESPOND"
+// @Failure 401 {object} utils.ErrorRespond "UNAUTHORIZED RESPOND"
+// @Failure 500 {object} utils.ErrorRespond "INTERNAL SERVER ERROR RESPOND"
+// @Router /api/v1/units [GET]
 func (handler attributeUnitHandler) fetch(ctx *gin.Context) {
 	data, err := handler.service.UnitList(ctx)
 	if err != nil {
@@ -21,6 +32,18 @@ func (handler attributeUnitHandler) fetch(ctx *gin.Context) {
 	utils.NewHTTPRespond(ctx, http.StatusOK, data)
 }
 
+// units godoc
+// @Schemes
+// @Summary Store Unit Data
+// @Description Create new Unit.
+// @Tags Attribute Units
+// @Accept json
+// @Produce json
+// @Param body formData UnitForm true "body"
+// @Failure 401 {object} utils.ErrorRespond "UNAUTHORIZED RESPOND"
+// @Failure 422 {object} utils.ValidationErrorRespond "UNPROCESSABLE ENTITY RESPOND"
+// @Failure 500 {object} utils.ErrorRespond "INTERNAL SERVER ERROR RESPOND"
+// @Router /api/v1/units [POST]
 func (handler attributeUnitHandler) add(ctx *gin.Context) {
 	var form UnitForm
 
@@ -44,6 +67,20 @@ func (handler attributeUnitHandler) add(ctx *gin.Context) {
 	utils.NewHTTPRespond(ctx, http.StatusCreated, nil)
 }
 
+// units godoc
+// @Schemes
+// @Summary Update Unit Data
+// @Description Update Unit Data by ID.
+// @Tags Attribute Units
+// @Accept json
+// @Produce json
+// @Param id   			path     int  	true "unit id"
+// @Param body formData UnitForm true "body"
+// @Failure 400 {object} utils.ErrorRespond "BAD REQUEST RESPOND"
+// @Failure 401 {object} utils.ErrorRespond "UNAUTHORIZED RESPOND"
+// @Failure 422 {object} utils.ValidationErrorRespond "UNPROCESSABLE ENTITY RESPOND"
+// @Failure 500 {object} utils.ErrorRespond "INTERNAL SERVER ERROR RESPOND"
+// @Router /api/v1/units/{id} [PATCH]
 func (handler attributeUnitHandler) edit(ctx *gin.Context) {
 	id, ok := utils.GetIDParam(ctx, "id")
 	if !ok {
@@ -73,6 +110,19 @@ func (handler attributeUnitHandler) edit(ctx *gin.Context) {
 	utils.NewHTTPRespond(ctx, http.StatusOK, nil)
 }
 
+// units godoc
+// @Schemes
+// @Summary Delete Unit Data
+// @Description Delete Unit Data by ID.
+// @Tags Attribute Units
+// @Accept json
+// @Produce json
+// @Param id path int true "unit id"
+// @Success 204 "NO CONTENT RESPOND"
+// @Failure 400 {object} utils.ErrorRespond "BAD REQUEST RESPOND"
+// @Failure 401 {object} utils.ErrorRespond "UNAUTHORIZED RESPOND"
+// @Failure 500 {object} utils.ErrorRespond "INTERNAL SERVER ERROR RESPOND"
+// @Router /api/v1/units/{id} [DELETE]
 func (handler attributeUnitHandler) destroy(ctx *gin.Context) {
 	id, ok := utils.GetIDParam(ctx, "id")
 	if !ok {
