+++ internal/order/provider.go
@@ -1,0 +1,6 @@
+package order
+
+import "github.com/gin-gonic/gin"
+
+func New(router *gin.RouterGroup) {
+}
