--- internal/transaction/handler/http/order_handler.go
@@ -1,6 +1,0 @@
-package http
-
-// type orderHandler struct{}
-// add order items to transaction
-// payment
-// func NewOrderHandler() {}
