--- internal/transaction/handler/http/transaction_handler.go
@@ -1,13 +1,0 @@
-package http
-
-// type  transactioHandler struct{}
-
-// list
-// detail
-// create
-// update
-//
-// we will not impl delete func
-// user can only update the transaction data
-
-// func NewTransactionHandler() {}
