--- internal/transaction/repository/sql/transaction_sql_repository.go
@@ -1,1 +1,0 @@
-package sql
