--- internal/transaction/service/order_service.go
@@ -1,1 +1,0 @@
-package service
