--- internal/transaction/service/transaction_service.go
@@ -1,1 +1,0 @@
-package service
