+++ misc/conf/.env.example
@@ -1,8 +1,8 @@
 APP_NAME="POSBE"
 APP_PORT=":8000"
 APP_DEBUG=true
-APP_VERSION="0.0.1-dev"
-API_LIMITER="500-M"
+APP_VERSION="0.2.0"
+API_LIMITER="250-M"
 
 POSTGRES_DSN_URL="postgresql://postgres:@127.0.0.1:5432/posbe?sslmode=disable"
 REDIS_DSN_URL="localhost:6379"
